Extract composer text field measurement hooks
